+++ hipaycard-cmp/src/iosMain/kotlin/com/hipay/card/cmp/HiPayCardEntry.ios.kt
@@ -8,6 +8,8 @@ import com.hipay.card.style.HiPayCardEntryStyle
 import com.hipay.card.validation.CardNetwork
 import com.hipay.core.HiPayConfig
 import com.hipay.core.gateway.model.CustomerInfo
+import com.hipay.card.PaymentPhase
+import com.hipay.core.gateway.model.OrderOptions
 import com.hipay.core.gateway.model.Transaction
 
 /**
@@ -61,6 +63,7 @@ actual class HiPayCardController actual constructor(
         shipping: CustomerInfo?,
         threeDS: HiPayThreeDSMode,
         saveCard: Boolean,
+        options: OrderOptions?,
     ): Transaction = impl.pay(
         orderId = orderId,
         amount = amount,
@@ -74,6 +77,7 @@ actual class HiPayCardController actual constructor(
         shipping = shipping,
         threeDS = threeDS,
         saveCard = saveCard,
+        options = options,
     )
 
     actual suspend fun payWithSavedCard(
@@ -89,6 +93,7 @@ actual class HiPayCardController actual constructor(
         customer: CustomerInfo?,
         shipping: CustomerInfo?,
         threeDS: HiPayThreeDSMode,
+        options: OrderOptions?,
     ): Transaction = impl.payWithSavedCard(
         card = card,
         orderId = orderId,
@@ -102,7 +107,10 @@ actual class HiPayCardController actual constructor(
         customer = customer,
         shipping = shipping,
         threeDS = threeDS,
+        options = options,
     )
+
+    actual val paymentPhase: PaymentPhase? get() = impl.paymentPhase
 
     actual val lastSaveOutcome: SavedCardOutcome? get() = impl.lastSaveOutcome